# QUICKSTEP-69 Query optimization with ExactFilter