… query_optimizer/CMakeLists.txt
… query_optimizer/cost_model/CMakeLists.txt
… query_optimizer/cost_model/StarSchemaSimpleCostModel.cpp
… query_optimizer/cost_model/StarSchemaSimpleCostModel.hpp
… query_optimizer/physical/CMakeLists.txt
… query_optimizer/physical/FilterJoin.cpp
… query_optimizer/physical/FilterJoin.hpp
… query_optimizer/physical/LIPFilterConfiguration.hpp
… query_optimizer/physical/PatternMatcher.hpp
… query_optimizer/rules/AttachLIPFilters.cpp
… query_optimizer/rules/CMakeLists.txt
… query_optimizer/rules/InjectJoinFilters.cpp
… query_optimizer/rules/InjectJoinFilters.hpp
… query_optimizer/tests/OptimizerTextTest.cpp
… relational_operators/BuildLIPFilterOperator.cpp
… relational_operators/BuildLIPFilterOperator.hpp
… relational_operators/CMakeLists.txt
… relational_operators/WorkOrderFactory.cpp
… utility/CMakeLists.txt
… utility/PlanVisualizer.cpp
… utility/lip_filter/BitVectorExactFilter.hpp
… utility/lip_filter/CMakeLists.txt
… utility/lip_filter/LIPFilter.hpp
… utility/lip_filter/LIPFilter.proto
… utility/lip_filter/LIPFilterDeployment.cpp
… utility/lip_filter/LIPFilterDeployment.hpp
… utility/lip_filter/LIPFilterFactory.cpp