This is a follow-up optimization based on the facility provided by LIPFilters. Note that LIP (lookahead information passing) is an optimization that we can inject efficient filters (e.g. bloom filters) into Select/HashJoin/Aggregate operators to pre-filter the input relations.

This PR strength-reduces `HashJoin`s (including inner/semi/anti joins) into `FilterJoin`s. The semantics of a `FilterJoin` is simple: if certain conditions are met, we can build a bit vector from the build side and use the bit vector to _filter_ the probe side.

The execution part is slightly more optimized: a `FilterJoin` will not always be converted into a `SelectOperator` plus a `LIPFilter` as its semantics indicates. Instead, in most situations we can avoid creating the `SelectOperator` by attaching the `LIPFilter` properly to some downstream operators – thus avoid unnecessary materialization of intermediate relations.


Below shows the performance improvement for SSB scale factor 100 on a cloudlab machine:

**SSB SF100**|**master (ms)**|**w/ ExactFilter (ms)**
:-----:|:-----:|:-----:
Q01|709|574
Q02|648|593
Q03|605|564
Q04|906|675
Q05|754|457
Q06|498|549
Q07|1687|1696
Q08|598|591
Q09|481|470
Q10|450|442
Q11|1208|882
Q12|876|656
Q13|515|475
Total|9937|8625