@hbdeshmukh 2017-01-31
Looks good to me. Will merge after you address the comments and fix the conflict.